requirements.txt
altair
pandas
streamlit
openai
scikit_learn

torch
torchvision
Pillow
requests
